Feature/ad comp (#8)
* 広告コンポーネント仕様変更

* URP Sample Asset内の広告Prefabパラメータ設定

* Bug fix

* Change variable type

* Add advertisement prefabs for HDRP

* Add advertisement prefabs for URP

+++ PlateauToolkit.Sandbox/Runtime/PlateauSandboxAdvertisement.cs
@@ -24,8 +24,8 @@ namespace PlateauToolkit.Sandbox.Runtime
             Z,
         }
 
-        public int textureAspectWidth = 1;
-        public int textureAspectHeight = 3;
+        public float textureAspectWidth = 1;
+        public float textureAspectHeight = 3;
         public int targetMaterialNumber;
         public string targetTextureProperty = "_MainTex";
         public FrontAxis frontAxis;
